refact: changing LFU Long to Integer

+++ app/src/main/java/page/x/TLB/algoritmos/substituicao/LFU.java
@@ -12,7 +12,7 @@ public class LFU implements AlgoritmoSubstituicaoI {
     
     private int quantidadeEntries;
 
-    private List<Pair<Long, TlbEntry>> entries;
+    private List<Pair<Integer, TlbEntry>> entries;
 
     public LFU(int quantidadeEntries) {
         this.quantidadeEntries = quantidadeEntries;
@@ -21,7 +21,7 @@ public class LFU implements AlgoritmoSubstituicaoI {
 
     @Override
     public Long mapearPagina(Long vpn) throws MissInterruption {
-        for (Pair<Long, TlbEntry> pairAtual : entries) {
+        for (Pair<Integer, TlbEntry> pairAtual : entries) {
             TlbEntry entryAtual = pairAtual.getPair2();
             if (entryAtual.getVirtualPageNumber().equals(vpn)) {
                 this.adicionaAcesso(pairAtual);
@@ -31,15 +31,15 @@ public class LFU implements AlgoritmoSubstituicaoI {
         throw new MissInterruption();
     }
     
-    private void adicionaAcesso(Pair<Long, TlbEntry> pairAtual) {
-        Long adiciona = pairAtual.getPair1() + 1;
+    private void adicionaAcesso(Pair<Integer, TlbEntry> pairAtual) {
+        Integer adiciona = pairAtual.getPair1() + 1;
         pairAtual.setPair1(adiciona);
     }
 
     @Override
     public void addPaginaMapeada(Long vpn, Long pfn) {
         TlbEntry tlbEntry = new TlbEntry(vpn, pfn);
-        Pair<Long, TlbEntry> tlbEntryPair = new Pair<Long,TlbEntry>(1L, tlbEntry);
+        Pair<Integer, TlbEntry> tlbEntryPair = new Pair<Integer,TlbEntry>(1, tlbEntry);
         if (quantidadeEntries == entries.size()) {
             this.removeMenosAcesso();
         }
@@ -47,8 +47,8 @@ public class LFU implements AlgoritmoSubstituicaoI {
     }
 
     private void removeMenosAcesso() {
-        Pair<Long, TlbEntry> menorAcesso = entries.getFirst();
-        for (Pair<Long, TlbEntry> i : entries) {
+        Pair<Integer, TlbEntry> menorAcesso = entries.getFirst();
+        for (Pair<Integer, TlbEntry> i : entries) {
             if (i.getPair1() < menorAcesso.getPair1()) {
                 menorAcesso = i;
             }
